requirements.txt
apache-airflow-providers-amazon
requests
pytest
testcontainers==3.7.1